captcha/requirements.txt
requests==2.22.0
pytesseract==0.2.7
Pillow==6.1.0
selenium==3.141.0

